+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   `Dialog`, `Lightbox` & `PopoverDialog`: set default heading level context to `h2`
+-   `UserBlock`: added `additionalFields` and `after` props
 
 ## [3.11.3][] - 2025-03-04